refactor: extract hero-shader module
Moves the HeroNameShader class (raw-WebGL iridescent name shader for
the hero <h1>) out of js/main.js into the new js/hero-shader.js
module.  No behavioural change: main.js imports HeroNameShader and
re-exports it for the test suite.

This module does not depend on Three.js — it uses raw WebGL — so it
only imports `isLowPowerDevice` from utils.js.

207 tests still pass.
